Use int instead of long
Use int for the values instead of long.

+++ tests/AdventOfCode.Tests/Puzzles/Y2021/Day21Tests.cs
@@ -21,7 +21,7 @@ public sealed class Day21Tests : PuzzleTest
         };
 
         // Act
-        long actual = Day21.Play(players);
+        int actual = Day21.Play(players);
 
         // Assert
         actual.ShouldBe(739785);
